Adding new textures and adjusting NPC animations
New PNG textures for NPCs (`npc2` and `npc3`) have been added, which include the "idle" and "talk" states. These changes expand the visual possibilities of the game and add new animations for characters.

In addition, the following code modifications have been made:
- Changing the path to the character texture in the `LessonSprite.cpp` file.
- Adjusting the player and NPC properties, including increasing the number of animation frames and adjusting collision rectangles.
- Adding a second NPC (`npc2`) and integrating it into the main program loop.
- Changing the length of the NPC animation frame and adjusting the collision rectangle calculation.
- Adjusting the `Direction` enumeration and adding new states for players.

These changes improve the visual aspect of the game, add new features, and expand the possibilities of interacting with NPCs.